enable DAT for the default clickhouse destination
