+++ airbyte-integrations/connectors/destination-clickhouse/src/test-integration/java/io/airbyte/integrations/destination/clickhouse/ClickhouseTestDataComparator.java
@@ -1,0 +1,25 @@
+package io.airbyte.integrations.destination.clickhouse;
+
+import io.airbyte.integrations.destination.ExtendedNameTransformer;
+import io.airbyte.integrations.standardtest.destination.comparator.AdvancedTestDataComparator;
+
+import java.util.ArrayList;
+import java.util.List;
+
+public class ClickhouseTestDataComparator extends AdvancedTestDataComparator {
+
+    private final ExtendedNameTransformer namingResolver = new ExtendedNameTransformer();
+
+    @Override
+    protected List<String> resolveIdentifier(final String identifier) {
+        final List<String> result = new ArrayList<>();
+        final String resolved = namingResolver.getIdentifier(identifier);
+        result.add(identifier);
+        result.add(resolved);
+        if (!resolved.startsWith("\"")) {
+            result.add(resolved.toLowerCase());
+            result.add(resolved.toUpperCase());
+        }
+        return result;
+    }
+}
